+++ jasentool/cli.py
@@ -240,7 +240,8 @@ def concatenate_files_cmd(input_files, output_file):
 @click.option('--mykrobe', type=click.Path(), default=None)
 @click.option('--nanoplot', type=click.Path(), default=None)
 @click.option('--nextflow-run-info', type=click.Path(), default=None)
-@click.option('--postalnqc', type=click.Path(), default=None)
+@click.option('--samtools-bedcov', type=click.Path(), default=None)
+@click.option('--samtools-stats', type=click.Path(), default=None)
 @click.option('--quast', type=click.Path(), default=None)
 @click.option('--ref-genome-annotation', type=click.Path(), default=None)
 @click.option('--ref-genome-sequence', type=click.Path(), default=None)
@@ -263,11 +264,12 @@ def concatenate_files_cmd(input_files, output_file):
 @click.option('-o', '--output', required=True, type=click.Path())
 def create_yaml_cmd(amrfinder, bam, bai, chewbbaca, emmtyper, gambitcore, groups,
                     kleborate, kleborate_hamronization, kraken, lims_id, mlst,
-                    mykrobe, nanoplot, nextflow_run_info, postalnqc, quast,
+                    mykrobe, nanoplot, nextflow_run_info, quast,
                     ref_genome_annotation, ref_genome_sequence, resfinder,
-                    sample_id, sample_name, samtools, sccmec, serotypefinder,
-                    shigapass, ska_index, software_info, sourmash_signature,
-                    spatyper, tb_grading_rules_bed, tbdb_bed, tbprofiler,
+                    sample_id, sample_name, samtools, samtools_bedcov,
+                    samtools_stats, sccmec, serotypefinder, shigapass,
+                    ska_index, software_info, sourmash_signature, spatyper,
+                    tb_grading_rules_bed, tbdb_bed, tbprofiler,
                     vcf, virulencefinder, output):
     """Create YAML input file for Bonsai upload."""
     options = types.SimpleNamespace(
@@ -276,10 +278,11 @@ def create_yaml_cmd(amrfinder, bam, bai, chewbbaca, emmtyper, gambitcore, groups
         kleborate=kleborate, kleborate_hamronization=kleborate_hamronization,
         kraken=kraken, lims_id=lims_id, mlst=mlst, mykrobe=mykrobe,
         nanoplot=nanoplot, nextflow_run_info=nextflow_run_info,
-        postalnqc=postalnqc, quast=quast,
+        quast=quast,
         ref_genome_annotation=ref_genome_annotation,
         ref_genome_sequence=ref_genome_sequence, resfinder=resfinder,
         sample_id=sample_id, sample_name=sample_name, samtools=samtools,
+        samtools_bedcov=samtools_bedcov, samtools_stats=samtools_stats,
         sccmec=sccmec, serotypefinder=serotypefinder, shigapass=shigapass,
         ska_index=ska_index, software_info=software_info,
         sourmash_signature=sourmash_signature, spatyper=spatyper,